Refactor Match class into its own space
